20230420_16:11 Showed to Shaon, seems alright. Now proceeding to MLE for the two params - mu, sigma of gaussian
Note: I didn't understand what Shaon said about single age parameter inference. If its a single age, then what's the distribution for? The gaussian with mu and sigma correponds to an age distribution. If i take all ages as the same, then then there is not gaussian to be maximised over